+++ src/modules/dashboard/components/DecisionPreview/index.ts
@@ -1,1 +1,1 @@
-export { default } from './DecisionPreviewForm';
+export { default } from './DecisionPreview';
